+++ workers/ohsome_quality_analyst/topics/models.py
@@ -25,6 +25,7 @@ class TopicDefinition(BaseTopic):
     """Includes the ohsome API endpoint and parameters needed to retrieve the data."""
 
     endpoint: str
+    aggregation_type: str
     filter: str
     indicators: List[str]
     project: str