+++ workers/tests/unittests/test_topic.py
@@ -28,6 +28,7 @@ def test_topic_definition():
         indicators=["mapping-saturation"],
         project="core",
         endpoint="endpoint",
+        aggregation_type="aggregation_type",
         filter="filter",
     )
     TopicDefinition(
@@ -37,6 +38,7 @@ def test_topic_definition():
         indicators=["mapping-saturation"],
         project="core",
         endpoint="endpoint",
+        aggregation_type="aggregation_type",
         filter="filter",
         source="source",
     )
@@ -47,6 +49,7 @@ def test_topic_definition():
         indicators=["mapping-saturation", "minimal"],
         project="core",
         endpoint="endpoint",
+        aggregation_type="aggregation_type",
         filter="filter",
         source="source",
     )
@@ -57,6 +60,7 @@ def test_topic_definition():
         indicators=[],
         project="core",
         endpoint="endpoint",
+        aggregation_type="aggregation_type",
         filter="filter",
         source="source",
     )
@@ -67,6 +71,7 @@ def test_topic_definition():
         indicators=["mapping-saturation"],
         project="core",
         endpoint="endpoint",
+        aggregation_type="aggregation_type",
         filter="filter",
         source="source",
         ratio_filter="ration_filter",